split XNAT unittests into a sub-package
