+++ arcana/testing/base.py
@@ -4,10 +4,8 @@ import os.path as op
 import os
 from arcana.utils import makedirs
 import subprocess as sp
-import operator
 import shutil
 from unittest import TestCase
-from functools import reduce
 import errno
 import sys
 import json
@@ -23,8 +21,6 @@ from arcana.processor import LinearProcessor
 from arcana.environment import StaticEnvironment
 from arcana.exception import ArcanaError
 from arcana.exception import ArcanaUsageError
-from nipype.interfaces.base import (
-    traits, TraitedSpec, BaseInterface, isdefined)
 
 logger = logging.getLogger('arcana')
 logger.setLevel(logging.WARNING)
@@ -561,93 +557,3 @@ class DummyTestCase(BaseTestCase):
             print(message)
         else:
             print("Test successful")
-
-
-class TestTestCase(BaseTestCase):
-
-    def test_test(self):
-        pass
-
-
-class TestMathInputSpec(TraitedSpec):
-
-    x = traits.Either(traits.Float(), traits.File(exists=True),
-                      traits.List(traits.Float),
-                      traits.List(traits.File(exists=True)),
-                      desc='first arg')
-    y = traits.Either(traits.Float(), traits.File(exists=True),
-                      mandatory=False, desc='second arg')
-    op = traits.Str(mandatory=True, desc='operation')
-
-    z = traits.File(genfile=True, mandatory=False,
-                    desc="Name for output file")
-
-    as_file = traits.Bool(False, desc="Whether to write as a file",
-                          usedefault=True)
-
-
-class TestMathOutputSpec(TraitedSpec):
-
-    z = traits.Either(traits.Float(), traits.File(exists=True),
-                      'output')
-
-
-class TestMath(BaseInterface):
-    """
-    A basic interface to test out the pipeline infrastructure
-    """
-
-    input_spec = TestMathInputSpec
-    output_spec = TestMathOutputSpec
-
-    def _run_interface(self, runtime):
-        return runtime
-
-    def _list_outputs(self):
-        x = self.inputs.x
-        y = self.inputs.y
-        if isinstance(x, basestring):
-            x = self._load_file(x)
-        if isinstance(y, basestring):
-            y = self._load_file(y)
-        try:
-            oper = getattr(operator, self.inputs.op)
-        except:
-            raise
-        if isdefined(y):
-            z = oper(x, y)
-        elif isinstance(x, list):
-            if isinstance(x[0], basestring):
-                x = [self._load_file(u) for u in x]
-            z = reduce(oper, x)
-        else:
-            raise Exception(
-                "If 'y' is not provided then x needs to be list")
-        outputs = self.output_spec().get()
-        if self.inputs.as_file:
-            z_path = op.abspath(self._gen_z_fname())
-            with open(z_path, 'w') as f:
-                f.write(str(z))
-            outputs['z'] = z_path
-        else:
-            outputs['z'] = z
-        return outputs
-
-    def _gen_filename(self, name):
-        if name == 'z':
-            fname = self._gen_z_fname()
-        else:
-            assert False
-        return fname
-
-    def _gen_z_fname(self):
-        if isdefined(self.inputs.z):
-            fname = self.inputs.z
-        else:
-            fname = 'z.txt'
-        return fname
-
-    @classmethod
-    def _load_file(self, path):
-        with open(path) as f:
-            return float(f.read())
