+++ test/unittests/repository/xnat/test_source_sink.py
@@ -1,365 +1,23 @@
 from __future__ import absolute_import
 from builtins import range
-from builtins import object
-import string
-import random
 import os
 import os.path as op
 import shutil
-import tempfile
-import re
 import json
 import time
 import unittest
 from multiprocessing import Process
-from unittest import TestCase
-import xnat
-from arcana.testing import (
-    BaseTestCase, BaseMultiSubjectTestCase)
+from arcana.testing import BaseTestCase
 from nipype.pipeline import engine as pe
 from nipype.interfaces.utility import IdentityInterface
 from arcana.repository.xnat import XnatRepository
-from arcana.repository.directory import DirectoryRepository
-from arcana.study import Study, StudyMetaClass
 from arcana.processor import LinearProcessor
 from arcana.interfaces.repository import RepositorySource, RepositorySink
-from arcana.data import (
-    FilesetSelector, AcquiredFilesetSpec, FilesetSpec, FieldSpec)
-from arcana.data.file_format import FileFormat
+from arcana.data import FilesetSelector
 from arcana.utils import PATH_SUFFIX, JSON_ENCODING
-from arcana.exception import ArcanaError
 from arcana.data.file_format.standard import text_format
-from arcana.repository.tree import Tree, Subject, Session, Visit
-from arcana.data import Fileset
-import sys
-import logging
-from future.utils import with_metaclass
-# Import TestExistingPrereqs study to test it on XNAT
-sys.path.insert(0, op.join(op.dirname(__file__), '..'))
-import test_data  # @UnresolvedImport @IgnorePep8
-sys.path.pop(0)
-
-# Import TestExistingPrereqs study to test it on XNAT
-sys.path.insert(0, op.join(op.dirname(__file__), '..',
-                                'study'))
-import test_study  # @UnresolvedImport @IgnorePep8
-sys.path.pop(0)
-
-# Import test_local to run TestProjectInfo on XNAT using TestOnXnat mixin
-sys.path.insert(0, op.join(op.dirname(__file__)))
-import test_directory  # @UnresolvedImport @IgnorePep8
-sys.path.pop(0)
-
-
-logger = logging.getLogger('arcana')
-
-dicom_format = FileFormat(name='dicom', extension=None,
-                          directory=True, within_dir_exts=['.dcm'])
-
-try:
-    SERVER = os.environ['ARCANA_TEST_XNAT']
-except KeyError:
-    SERVER = None
-
-SKIP_ARGS = (SERVER is None,
-             "Skipping as ARCANA_TEST_XNAT env var not set")
-
-
-class DummyStudy(with_metaclass(StudyMetaClass, Study)):
-
-    add_data_specs = [
-        AcquiredFilesetSpec('source1', text_format),
-        AcquiredFilesetSpec('source2', text_format, optional=True),
-        AcquiredFilesetSpec('source3', text_format, optional=True),
-        AcquiredFilesetSpec('source4', text_format, optional=True),
-        FilesetSpec('sink1', text_format, 'dummy_pipeline'),
-        FilesetSpec('sink3', text_format, 'dummy_pipeline'),
-        FilesetSpec('sink4', text_format, 'dummy_pipeline'),
-        FilesetSpec('subject_sink', text_format, 'dummy_pipeline',
-                    frequency='per_subject'),
-        FilesetSpec('visit_sink', text_format, 'dummy_pipeline',
-                    frequency='per_visit'),
-        FilesetSpec('study_sink', text_format, 'dummy_pipeline',
-                    frequency='per_study'),
-        FilesetSpec('resink1', text_format, 'dummy_pipeline'),
-        FilesetSpec('resink2', text_format, 'dummy_pipeline'),
-        FilesetSpec('resink3', text_format, 'dummy_pipeline'),
-        FieldSpec('field1', int, 'dummy_pipeline'),
-        FieldSpec('field2', float, 'dummy_pipeline'),
-        FieldSpec('field3', str, 'dummy_pipeline')]
-
-    def dummy_pipeline(self, **name_maps):
-        return self.pipeline('dummy_pipeline', name_maps=name_maps)
-
-
-class TestStudy(with_metaclass(StudyMetaClass, Study)):
-
-    add_data_specs = [
-        AcquiredFilesetSpec('fileset1', text_format),
-        AcquiredFilesetSpec('fileset2', text_format, optional=True),
-        AcquiredFilesetSpec('fileset3', text_format),
-        AcquiredFilesetSpec('fileset5', text_format, optional=True)]
-
-
-def filter_scans(names):
-    return sorted(f for f in sorted(names)
-                  if (f != XnatRepository.PROV_SCAN and
-                      not f.endswith(XnatRepository.MD5_SUFFIX)))
-
-
-class CreateXnatProjectMixin(object):
-
-    PROJECT_NAME_LEN = 12
-
-    @property
-    def project(self):
-        """
-        Creates a random string of letters and numbers to be the
-        project ID
-        """
-        try:
-            return self._project
-        except AttributeError:
-            self._project = ''.join(
-                random.choice(string.ascii_uppercase + string.digits)
-                for _ in range(self.PROJECT_NAME_LEN))
-            return self._project
-
-    def _create_project(self, project_name=None):
-        if project_name is None:
-            project_name = self.project
-        if SERVER == 'https://mbi-xnat.erc.monash.edu.au':
-            raise ArcanaError(
-                "Shouldn't be creating projects on the production "
-                "server")
-        with xnat.connect(SERVER) as login:
-            uri = '/data/archive/projects/{}'.format(project_name)
-            query = {'xsiType': 'xnat:projectData', 'req_format': 'qa'}
-            response = login.put(uri, query=query)
-            if response.ok:
-                logger.info("Created test project '{}'"
-                            .format(project_name))
-
-    def _delete_project(self, project_name=None):
-        if project_name is None:
-            project_name = self.project
-        with xnat.connect(SERVER) as login:
-            login.projects[project_name].delete()
-
-
-class TestOnXnatMixin(CreateXnatProjectMixin):
-
-    def session_label(self, project=None, subject=None, visit=None,
-                      from_study=None):
-        if project is None:
-            project = self.project
-        if subject is None:
-            subject = self.SUBJECT
-        if visit is None:
-            visit = self.VISIT
-        label = '_'.join((project, subject, visit))
-        if from_study is not None:
-            label += '_' + from_study
-        return label
-
-    def session_cache(self, base_dir=None, project=None, subject=None,
-                      visit=None, from_study=None):
-        if base_dir is None:
-            base_dir = self.cache_dir
-        if project is None:
-            project = self.project
-        if subject is None:
-            subject = self.SUBJECT
-        return op.join(
-            base_dir, project, '{}_{}'.format(project, subject),
-            self.session_label(project=project, subject=subject,
-                               visit=visit, from_study=from_study))
-
-    def setUp(self):
-        BaseTestCase.setUp(self)
-        shutil.rmtree(self.cache_dir, ignore_errors=True)
-        os.makedirs(self.cache_dir)
-        self._create_project()
-        with self._connect() as login:
-            xproject = login.projects[self.project]
-            xsubject = login.classes.SubjectData(
-                label='{}_{}'.format(self.project, self.SUBJECT),
-                parent=xproject)
-            xsession = login.classes.MrSessionData(
-                label=self.session_label(),
-                parent=xsubject)
-            for fileset in self.session.filesets:
-                # Create fileset with an integer ID instead of the
-                # label
-                query = {'xsiType': 'xnat:mrScanData',
-                         'req_format': 'qa',
-                         'type': fileset.name}
-                uri = '{}/scans/{}'.format(xsession.fulluri, fileset.id)
-                login.put(uri, query=query)
-                # Get XnatPy object for newly created fileset
-                xfileset = login.classes.MrScanData(uri,
-                                                    xnat_session=login)
-                resource = xfileset.create_resource(
-                    fileset.format.name.upper())
-                if fileset.format.directory:
-                    for fname in os.listdir(fileset.path):
-                        resource.upload(
-                            op.join(fileset.path, fname), fname)
-                else:
-                    resource.upload(fileset.path, fileset.fname)
-            for field in self.session.fields:
-                xsession.fields[field.name] = field.value
-
-    def tearDown(self):
-        # Clean up working dirs
-        shutil.rmtree(self.cache_dir, ignore_errors=True)
-        # Clean up session created for unit-test
-        self._delete_project()
-
-    def _connect(self):
-        return xnat.connect(SERVER)
-
-
-class TestMultiSubjectOnXnatMixin(CreateXnatProjectMixin):
-
-    sanitize_id_re = re.compile(r'[^a-zA-Z_0-9]')
-
-    def setUp(self):
-        self._clean_up()
-        self._repository = XnatRepository(project_id=self.project,
-                                          server=SERVER,
-                                          cache_dir=self.cache_dir)
-        self.BASE_CLASS.setUp(self)
-        local_repository = DirectoryRepository(self.project_dir)
-        tree = local_repository.tree()
-        self._create_project()
-        repo = XnatRepository(SERVER, self.project, '/tmp')
-        with repo:
-            for node in tree:
-                for fileset in node.filesets:
-                    repo.put_fileset(fileset)
-                for field in node.fields:
-                    repo.put_field(field)
-
-    def _upload_datset(self, xnat_login, fileset, xsession):
-        if self._is_derived(fileset):
-            type_name = self._derived_name(fileset)
-        else:
-            type_name = fileset.name
-        xfileset = xnat_login.classes.MrScanData(
-            type=type_name, parent=xsession)
-        xresource = xfileset.create_resource(
-            fileset.format.name.upper())
-        if fileset.format.directory:
-            for fname in os.listdir(fileset.path):
-                fpath = op.join(fileset.path, fname)
-                xresource.upload(fpath, fname)
-        else:
-            if not op.exists(fileset.path):
-                raise ArcanaError(
-                    "Cannot upload fileset {} as path ({}) does "
-                    "not exist".format(fileset, fpath))
-            xresource.upload(
-                fileset.path,
-                op.basename(fileset.path))
-
-    @classmethod
-    def _is_derived(cls, fileset):
-        # return fileset.name.endswith(self.DERIVED_SUFFIX
-        return '_' in fileset.name
-
-    @classmethod
-    def _derived_name(cls, fileset):
-        # return name[:-len(self.DERIVED_SUFFIX)]
-        return fileset.name
-
-    def tearDown(self):
-        self._clean_up()
-        self._delete_project()
-
-    def _clean_up(self):
-        # Clean up working dirs
-        shutil.rmtree(self.cache_dir, ignore_errors=True)
-
-    @property
-    def repository(self):
-        return self._repository
-
-    @property
-    def xnat_session_name(self):
-        return '{}_{}'.format(self.project, self.base_name)
-
-    @property
-    def project_dir(self):
-        return op.join(self.repository_path, self.base_name)
-
-    @property
-    def output_cache_dir(self):
-        return self._output_cache_dir
-
-    @property
-    def base_name(self):
-        return self.BASE_CLASS._get_name()
-
-    def _full_subject_id(self, subject):
-        return self.project + '_' + subject
-
-    def get_session_dir(self, subject=None, visit=None,
-                        frequency='per_session'):
-        if subject is None and frequency in ('per_session', 'per_subject'):
-            subject = self.SUBJECT
-        if visit is None and frequency in ('per_session', 'per_visit'):
-            visit = self.VISIT
-        if frequency == 'per_session':
-            assert subject is not None
-            assert visit is not None
-            parts = [self.project, subject, visit]
-        elif frequency == 'per_subject':
-            assert subject is not None
-            assert visit is None
-            parts = [self.project, subject, XnatRepository.SUMMARY_NAME]
-        elif frequency == 'per_visit':
-            assert visit is not None
-            assert subject is None
-            parts = [self.project, XnatRepository.SUMMARY_NAME, visit]
-        elif frequency == 'per_study':
-            assert subject is None
-            assert visit is None
-            parts = [self.project, XnatRepository.SUMMARY_NAME,
-                     XnatRepository.SUMMARY_NAME]
-        else:
-            assert False
-        session_id = '_'.join(parts)
-        session_path = op.join(self.output_cache_dir, session_id)
-        if not op.exists(session_path):
-            raise ArcanaError(
-                "Session path '{}' does not exist".format(session_path))
-        return session_path
-
-    def output_file_path(self, fname, from_study, subject=None, visit=None,
-                         frequency='per_session'):
-        try:
-            acq_path = self.BASE_CLASS.output_file_path(
-                self, fname, from_study, subject=subject, visit=visit,
-                frequency=frequency, derived=False)
-        except KeyError:
-            acq_path = None
-        try:
-            proc_path = self.BASE_CLASS.output_file_path(
-                self, fname, from_study, subject=subject, visit=visit,
-                frequency=frequency, derived=True)
-        except KeyError:
-            proc_path = None
-        if acq_path is not None and op.exists(acq_path):
-            if op.exists(proc_path):
-                raise ArcanaError(
-                    "Both acquired and derived paths were found for "
-                    "'{}_{}' ({} and {})".format(from_study, fname, acq_path,
-                                                 proc_path))
-            path = acq_path
-        else:
-            path = proc_path
-        return path
+from arcana.testing.xnat import (
+    TestOnXnatMixin, SERVER, SKIP_ARGS, DummyStudy, filter_scans, logger)
 
 
 class TestXnatSourceAndSinkBase(TestOnXnatMixin, BaseTestCase):
@@ -852,210 +510,3 @@ class TestXnatSummarySourceAndSink(TestXnatSourceAndSinkBase):
                         from_study=self.SUMMARY_STUDY_NAME)].scans.keys())
             self.assertEqual(sorted(resinked_fileset_names),
                              ['resink1', 'resink2', 'resink3'])
-
-
-class TestDicomTagMatchAndIDOnXnat(TestOnXnatMixin,
-                                   test_data.TestDicomTagMatch):
-
-    BASE_CLASS = test_data.TestDicomTagMatch
-
-    @property
-    def ref_dir(self):
-        return op.join(
-            self.ref_path, self._get_name(self.BASE_CLASS))
-
-    def setUp(self):
-        test_data.TestDicomTagMatch.setUp(self)
-        TestOnXnatMixin.setUp(self)
-        # Set up DICOM headers
-        with xnat.connect(SERVER) as login:
-            xsess = login.projects[self.project].experiments[
-                '_'.join((self.project, self.SUBJECT, self.VISIT))]
-            login.put('/data/experiments/{}?pullDataFromHeaders=true'
-                      .format(xsess.id))
-
-    def tearDown(self):
-        TestOnXnatMixin.tearDown(self)
-        test_data.TestDicomTagMatch.tearDown(self)
-
-    @unittest.skipIf(*SKIP_ARGS)
-    def test_dicom_match(self):
-        study = test_data.TestMatchStudy(
-            name='test_dicom',
-            repository=XnatRepository(
-                project_id=self.project,
-                server=SERVER, cache_dir=tempfile.mkdtemp()),
-            processor=LinearProcessor(self.work_dir),
-            inputs=test_data.TestDicomTagMatch.DICOM_MATCH)
-        phase = list(study.data('gre_phase'))[0]
-        mag = list(study.data('gre_mag'))[0]
-        self.assertEqual(phase.name, 'gre_field_mapping_3mm_phase')
-        self.assertEqual(mag.name, 'gre_field_mapping_3mm_mag')
-
-    @unittest.skipIf(*SKIP_ARGS)
-    def test_id_match(self):
-        study = test_data.TestMatchStudy(
-            name='test_dicom',
-            repository=XnatRepository(
-                project_id=self.project,
-                server=SERVER, cache_dir=tempfile.mkdtemp()),
-            processor=LinearProcessor(self.work_dir),
-            inputs=[
-                FilesetSelector('gre_phase', dicom_format, id=7),
-                FilesetSelector('gre_mag', dicom_format, id=6)])
-        phase = list(study.data('gre_phase'))[0]
-        mag = list(study.data('gre_mag'))[0]
-        self.assertEqual(phase.name, 'gre_field_mapping_3mm_phase')
-        self.assertEqual(mag.name, 'gre_field_mapping_3mm_mag')
-
-    @unittest.skipIf(*SKIP_ARGS)
-    def test_order_match(self):
-        test_data.TestDicomTagMatch.test_order_match(self)
-
-
-class TestFilesetCacheOnPathAccess(TestOnXnatMixin,
-                                   BaseTestCase):
-
-    INPUT_DATASETS = {'fileset': '1'}
-
-    @unittest.skipIf(*SKIP_ARGS)
-    def test_cache_on_path_access(self):
-        tmp_dir = tempfile.mkdtemp()
-        repository = XnatRepository(
-            project_id=self.project,
-            server=SERVER, cache_dir=tmp_dir)
-        tree = repository.tree(
-            subject_ids=[self.SUBJECT],
-            visit_ids=[self.VISIT])
-        # Get a fileset
-        fileset = list(list(list(tree.subjects)[0].sessions)[0].filesets)[0]
-        self.assertEqual(fileset._path, None)
-        target_path = op.join(
-            tmp_dir, self.project,
-            '{}_{}'.format(self.project, self.SUBJECT),
-            '{}_{}_{}'.format(self.project, self.SUBJECT, self.VISIT),
-            fileset.fname)
-        # This should implicitly download the fileset
-        self.assertEqual(fileset.path, target_path)
-        with open(target_path) as f:
-            self.assertEqual(f.read(),
-                             self.INPUT_DATASETS[fileset.name])
-
-
-class TestExistingPrereqsOnXnat(TestMultiSubjectOnXnatMixin,
-                                test_study.TestExistingPrereqs):
-
-    BASE_CLASS = test_study.TestExistingPrereqs
-
-    @unittest.skipIf(*SKIP_ARGS)
-    def test_per_session_prereqs(self):
-        super(TestExistingPrereqsOnXnat, self).test_per_session_prereqs()
-
-
-class TestXnatCache(TestMultiSubjectOnXnatMixin,
-                    BaseMultiSubjectTestCase):
-
-    BASE_CLASS = BaseMultiSubjectTestCase
-    STRUCTURE = {
-        'subject1': {
-            'visit1': ['fileset1', 'fileset2', 'fileset3'],
-            'visit2': ['fileset1', 'fileset2', 'fileset3']},
-        'subject2': {
-            'visit1': ['fileset1', 'fileset2', 'fileset3'],
-            'visit2': ['fileset1', 'fileset2', 'fileset3']}}
-
-    DATASET_CONTENTS = {'fileset1': 1,
-                        'fileset2': 2,
-                        'fileset3': 3}
-
-    @property
-    def input_tree(self):
-        sessions = []
-        visit_ids = set()
-        for subj_id, visits in list(self.STRUCTURE.items()):
-            for visit_id, filesets in list(visits.items()):
-                sessions.append(Session(subj_id, visit_id, filesets=[
-                    Fileset(d, text_format, subject_id=subj_id,
-                            visit_id=visit_id) for d in filesets]))
-                visit_ids.add(visit_id)
-        subjects = [Subject(i, sessions=[s for s in sessions
-                                         if s.subject_id == i])
-                    for i in self.STRUCTURE]
-        visits = [Visit(i, sessions=[s for s in sessions
-                                     if s.visit == i])
-                  for i in visit_ids]
-        return Tree(subjects=subjects, visits=visits)
-
-    @unittest.skipIf(*SKIP_ARGS)
-    def test_cache_download(self):
-        repository = XnatRepository(
-            project_id=self.project,
-            server=SERVER,
-            cache_dir=tempfile.mkdtemp())
-        study = self.create_study(
-            TestStudy, 'cache_download',
-            inputs=[
-                FilesetSelector('fileset1', text_format, 'fileset1'),
-                FilesetSelector('fileset3', text_format, 'fileset3')],
-            repository=repository)
-        study.cache_inputs()
-        for subject_id, visits in list(self.STRUCTURE.items()):
-            subj_dir = op.join(
-                repository.cache_dir, self.project,
-                '{}_{}'.format(self.project, subject_id))
-            for visit_id in visits:
-                sess_dir = op.join(
-                    subj_dir,
-                    '{}_{}_{}'.format(self.project, subject_id,
-                                      visit_id))
-                for inpt in study.inputs:
-                    self.assertTrue(op.exists(op.join(
-                        sess_dir, inpt.name + inpt.format.extension)))
-
-    @property
-    def base_name(self):
-        return self.name
-
-
-class TestProjectInfo(TestMultiSubjectOnXnatMixin,
-                      test_directory.TestDirectoryProjectInfo):
-
-    BASE_CLASS = test_directory.TestDirectoryProjectInfo
-
-    @unittest.skipIf(*SKIP_ARGS)
-    def test_project_info(self):
-        tree = self.repository.tree()
-        ref_tree = self.get_tree(self.repository, set_ids=True)
-        self.assertEqual(
-            tree, ref_tree,
-            "Generated project doesn't match reference:{}"
-            .format(tree.find_mismatch(ref_tree)))
-
-
-class TestConnectDisconnect(TestCase):
-
-    @unittest.skipIf(*SKIP_ARGS)
-    def test_connect_disconnect(self):
-        repository = XnatRepository(project_id='dummy',
-                                    server=SERVER,
-                                    cache_dir=tempfile.mkdtemp())
-        with repository:
-            self._test_open(repository)
-        self._test_closed(repository)
-
-        with repository:
-            self._test_open(repository)
-            with repository:
-                self._test_open(repository)
-            self._test_open(repository)
-        self._test_closed(repository)
-
-    def _test_open(self, repository):
-        repository._login.classes  # check connection
-
-    def _test_closed(self, repository):
-        self.assertRaises(
-            AttributeError,
-            getattr,
-            repository._login,
-            'classes')
